moved interfaces to common, update accordingly
